refactor remaining types
