+++ src/zos/types/mod.rs
@@ -1,1 +1,3 @@
 pub mod net;
+pub mod stats;
+pub mod version;
